Renamed to ReportCsvWriter
